Move IsolateContextSource alongside WorkNodeCodec
